Issues 255, 256, 257-chatInput with styled component
